# developer-birth-chart/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Developer Birth Chart API
  description: |
    Comprehensive API for the Personalized Developer Birth Chart project.

    This API enables:
    - GitHub profile analysis with astrological-style personality insights
    - Social media sharing and viral marketing
    - Team compatibility and dynamics analysis
    - Marketplace features with Stripe Connect
    - Real-time analytics and monitoring

    ## Authentication

    Most endpoints require authentication via GitHub OAuth or API keys. Include your token in the Authorization header:

    ```
    Authorization: Bearer <your-token>
    ```

    ## Rate Limiting

    - **Free tier**: 100 requests per hour
    - **Premium tier**: 1000 requests per hour
    - **Enterprise tier**: 5000 requests per hour

    Rate limit headers are included in all responses.
  version: 1.0.0
  contact:
    name: Developer Birth Chart Support
    email: support@developerbirthchart.com
    url: https://developerbirthchart.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.developerbirthchart.com/v1
    description: Production server
  - url: https://staging-api.developerbirthchart.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /subscriptions/create:
    post:
      tags:
        - Subscriptions
      summary: Create Subscription
      description: Create a new subscription for premium features.
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '200':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: Payment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    SubscriptionPlan:
      type: object
      properties:
        id:
          type: string
          description: Plan ID
          example: "premium_monthly"
        name:
          type: string
          description: Plan name
          example: "Premium Monthly"
        description:
          type: string
          description: Plan description
          example: "Advanced analytics and team features"
        price:
          type: number
          description: Price in cents
          example: 2900
        currency:
          type: string
          description: Currency code
          example: "usd"
        interval:
          type: string
          enum: [month, year]
          description: Billing interval
          example: "month"
        features:
          type: array
          items:
            type: string
          description: Plan features
          example: ["Unlimited charts", "Team analysis", "API access", "Priority support"]
        trialPeriodDays:
          type: integer
          description: Trial period in days
          example: 14
        isActive:
          type: boolean
          description: Whether plan is currently active
          example: true
    CreateSubscriptionRequest:
      type: object
      required:
        - planId
        - paymentMethodId
      properties:
        planId:
          type: string
          description: Plan ID to subscribe to
          example: "premium_monthly"
        paymentMethodId:
          type: string
          description: Payment method ID
          example: "pm_1234567890"
        customerId:
          type: string
          description: Customer ID (optional for new customers)
        trialPeriodDays:
          type: integer
          description: Custom trial period days
          minimum: 0
          maximum: 90
          example: 30
    SubscriptionResponse:
      type: object
      properties:
        subscriptionId:
          type: string
          description: Subscription ID
          example: "sub_1234567890"
        status:
          type: string
          enum: [active, trialing, past_due, canceled, unpaid]
          description: Subscription status
          example: "active"
        currentPeriodEnd:
          type: string
          format: date-time
          description: Current period end date
        cancelAtPeriodEnd:
          type: boolean
          description: Whether subscription will cancel at period end
        plan:
          $ref: '#/components/schemas/SubscriptionPlan'
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Error message
          example: "Invalid username provided"
        code:
          type: string
          description: Error code
          example: "INVALID_USERNAME"
        details:
          type: object
          description: Additional error details
          example:
            field: "username"
            reason: "Username contains invalid characters"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        requestId:
          type: string
          description: Request ID for tracing
          example: "req_1234567890"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
...